We now process all pipelines in the Go Server
We no longer process only the pipelines in the configuration but all of them in the Server. After breaking the production Go server enough number of times I've realized that people with 50+ pipelines are most likely to screw up the dependencies in a really really crazy manner. Also when you're trying this out, it is very hard to predict at one go what all piplines you need. We are most likely to miss something and screw things up.

The configuration is going forward more for overriding the default behaviour of willing to have more than 2 successful runs of every pipeline, also this includes the version that is a downstream dependency.